Update command to return int32s

+++ builtin/myplugin/command/command.go
@@ -58,15 +58,15 @@ func (c *Command) ExecuteFunc(cliArgs []string) interface{} {
 	return c.Execute
 }
 
-func (c *Command) ExecuteInfo(trm terminal.UI, env plugincore.Project) int64 {
+func (c *Command) ExecuteInfo(trm terminal.UI, env plugincore.Project) int32 {
 	return (&Info{Command: c}).Execute(trm, env)
 }
 
-func (c *Command) ExecuteDoThing(trm terminal.UI, flags map[string]interface{}) int64 {
+func (c *Command) ExecuteDoThing(trm terminal.UI, flags map[string]interface{}) int32 {
 	return (&DoThing{Command: c}).Execute(trm, flags)
 }
 
-func (c *Command) ExecuteInteractive(trm terminal.UI, flags map[string]interface{}) int64 {
+func (c *Command) ExecuteInteractive(trm terminal.UI, flags map[string]interface{}) int32 {
 	return (&Interactive{Command: c}).Execute(trm)
 }
 
@@ -102,7 +102,7 @@ func (c *Command) Flags() []*option.Option {
 	return []*option.Option{stringflag}
 }
 
-func (c *Command) Execute(trm terminal.UI, flags map[string]interface{}) int64 {
+func (c *Command) Execute(trm terminal.UI, flags map[string]interface{}) int32 {
 	trm.Output("You gave me the flag: " + flags["hehe"].(string))
 
 	trm.Output(c.Help())
