+++ builtin/myplugin/command/dothing.go
@@ -74,7 +74,7 @@ func (c *DoThing) Flags() []*option.Option {
 	return []*option.Option{booltest, stringflag}
 }
 
-func (c *DoThing) Execute(trm terminal.UI, flags map[string]interface{}) int64 {
+func (c *DoThing) Execute(trm terminal.UI, flags map[string]interface{}) int32 {
 	trm.Output("Tricked ya! I actually do nothing :P")
 	trm.Output("You gave me the stringflag: " + flags["stringflag"].(string))
 	return 0
